7 shaped binary array operation by bsxfun from matlab but with "hinge" shape

7 shaped binary array operation by bsxfun from matlab but with "hinge" shape

bsxfun_7_hinge_pow(A, B)

Arguments

A
matrix of size d1 by d2

B
matrix of size d3 by d2

Value

a matrix of dimension d3 by d1 by d2

Examples

bsxfun_7_hinge_pow(matrix(c(1,2,3,4,5,6),nrow=3,byrow=TRUE),
matrix(c(1:18),nrow=9,byrow=TRUE))
#> , , 1
#> 
#>       [,1]      [,2]         [,3]
#>  [1,]    1         3            5
#>  [2,]    1        27          125
#>  [3,]    1       243         3125
#>  [4,]    1      2187        78125
#>  [5,]    1     19683      1953125
#>  [6,]    1    177147     48828125
#>  [7,]    1   1594323   1220703125
#>  [8,]    1  14348907  30517578125
#>  [9,]    1 129140163 762939453125
#> 
#> , , 2
#> 
#>         [,1]        [,2]         [,3]
#>  [1,]      4          16 3.600000e+01
#>  [2,]     16         256 1.296000e+03
#>  [3,]     64        4096 4.665600e+04
#>  [4,]    256       65536 1.679616e+06
#>  [5,]   1024     1048576 6.046618e+07
#>  [6,]   4096    16777216 2.176782e+09
#>  [7,]  16384   268435456 7.836416e+10
#>  [8,]  65536  4294967296 2.821110e+12
#>  [9,] 262144 68719476736 1.015600e+14
#>