7 shaped binary array operation by bsxfun from matlab but with "hinge" shape; multiple along dimension d2

7 shaped binary array operation by bsxfun from matlab but with "hinge" shape; multiple along dimension d2

bsxfun_7_hinge_pow_prod(A, B)

Arguments

A
matrix of size d1 by d2

B
matrix of size d3 by d2

Value

a matrix of dimension d3 by d1 by d2

Examples

bsxfun_7_hinge_pow_prod(matrix(c(1,2,3,4,5,6),nrow=3,byrow=TRUE),
matrix(c(1:18),nrow=9,byrow=TRUE))
#>         [,1]         [,2]         [,3]
#>  [1,]      4 4.800000e+01 1.800000e+02
#>  [2,]     16 6.912000e+03 1.620000e+05
#>  [3,]     64 9.953280e+05 1.458000e+08
#>  [4,]    256 1.433272e+08 1.312200e+11
#>  [5,]   1024 2.063912e+10 1.180980e+14
#>  [6,]   4096 2.972033e+12 1.062882e+17
#>  [7,]  16384 4.279728e+14 9.565938e+19
#>  [8,]  65536 6.162809e+16 8.609344e+22
#>  [9,] 262144 8.874444e+18 7.748410e+25