generate second-level responses with taxonomy under DINA

generate second-level responses with taxonomy under DINA

generate_X_tax2(X1_sub, A_set20, D_mat, p_short, Q2, c, g, model = "DINA")

Arguments

X1_sub
The subset of individuals who has one particular level 1 latent attributes

A_set20
The true leve 2 set of latent attributes

D_mat
J1 by J2 matrix for indicating the tree structure

p_short
level 2 sparse population level prevalences

Q2
level 2 Q in a subset of level 1 latent attributes

c
1-slipping parameters

g
guess parameters

model
"DINA" (default); "DINO"

Value

X2 level 2 observed responses

Z latent attribute profiles for all individuals at level 2

X_true2 Ideal response matrix