PEM FOR SHRINKAGE - this is to take a candidate latent attribute set

PEM FOR SHRINKAGE - this is to take a candidate latent attribute set

get_PEM(X, Q, A, lambda, c_ini, g_ini, nu_ini, thres_c, model = "DINA")

Arguments

X
N by J data matrix

Q
J by K structural matrix

A
K-column latent attributes; candidates to be selected!

lambda
negative penalty parameter

c_ini
initial 1-slipping parameters

g_ini
initial guessing parameters

nu_ini
initial potentials

thres_c
threshold for making sure M step does not generate negative probabilities (algorithm 1 of Xu and Gu JMLR 2019)

model
"DINA" (default) or "DINO"

Value

nu vector of potential ("Delta")

c 1-slipping parameter

g guessing parameter